Rename ghost_hierarchy -> ui_tree
