+++ crates/bevy_ui/src/lib.rs
@@ -26,20 +26,20 @@ use bevy_reflect::{std_traits::ReflectDefault, Reflect};
 mod accessibility;
 mod focus;
 mod geometry;
-mod ghost_hierarchy;
 mod layout;
 mod render;
 mod stack;
 mod ui_node;
+mod ui_tree;
 
 pub use focus::*;
 pub use geometry::*;
-pub use ghost_hierarchy::*;
 pub use layout::*;
 pub use measurement::*;
 pub use render::*;
 pub use ui_material::*;
 pub use ui_node::*;
+pub use ui_tree::*;
 use widget::UiImageSize;
 
 /// The UI prelude.
